refactor:refactor client details screen to compose with multi module
