+++ core/common/src/main/java/com/mifos/core/common/utils/BaseUrl.kt
@@ -6,7 +6,7 @@ object BaseUrl {
 
     const val PROTOCOL_HTTPS = "https://"
 
-    const val API_ENDPOINT = "gsoc.mifos.community"
+    const val API_ENDPOINT = "demo.mifos.community"
 
     const val API_PATH = "/fineract-provider/api/v1/"
