--- core/designsystem/src/main/java/com/mifos/core/designsystem/component/MifosDialogBox.kt
@@ -1,125 +1,0 @@
-package com.mifos.core.designsystem.component
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.MutableState
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.Font
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.font.FontStyle
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.compose.ui.window.Dialog
-import androidx.compose.ui.window.DialogProperties
-import com.mifos.core.designsystem.R
-import com.mifos.core.designsystem.theme.BlueSecondary
-import com.mifos.core.designsystem.theme.White
-
-@Composable
-fun MifosSelectImageDialog(
-    showSelectImageDialog: MutableState<Boolean>,
-    takeImage: () -> Unit,
-    uploadImage: () -> Unit,
-    deleteImage: () -> Unit
-) {
-
-    Dialog(
-        onDismissRequest = { showSelectImageDialog.value = !showSelectImageDialog.value },
-        properties = DialogProperties(
-            dismissOnBackPress = true,
-            dismissOnClickOutside = true
-        )
-    ) {
-        Card(
-            colors = CardDefaults.cardColors(White),
-            shape = RoundedCornerShape(20.dp)
-        ) {
-            Column(
-                modifier = Modifier
-                    .padding(30.dp),
-                verticalArrangement = Arrangement.Center,
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                Text(
-                    text = "Please Select", modifier = Modifier.fillMaxWidth(), style = TextStyle(
-                        fontSize = 18.sp,
-                        fontWeight = FontWeight.Normal,
-                        fontStyle = FontStyle.Normal,
-                        fontFamily = FontFamily(Font(R.font.outfit_medium))
-                    ),
-                    color = Color.Black,
-                    textAlign = TextAlign.Center
-                )
-                Spacer(modifier = Modifier.height(20.dp))
-
-                Button(
-                    onClick = { takeImage() },
-                    colors = ButtonDefaults.buttonColors(BlueSecondary)
-                ) {
-                    Text(
-                        text = "Take new image",
-                        modifier = Modifier.fillMaxWidth(),
-                        style = TextStyle(
-                            fontSize = 18.sp,
-                            fontWeight = FontWeight.Normal,
-                            fontStyle = FontStyle.Normal,
-                            fontFamily = FontFamily(Font(R.font.outfit_medium))
-                        ),
-                        color = Color.Black,
-                        textAlign = TextAlign.Center
-                    )
-                }
-                Button(
-                    onClick = { uploadImage() },
-                    colors = ButtonDefaults.buttonColors(BlueSecondary)
-                ) {
-                    Text(
-                        text = "Upload new image",
-                        modifier = Modifier.fillMaxWidth(),
-                        style = TextStyle(
-                            fontSize = 18.sp,
-                            fontWeight = FontWeight.Normal,
-                            fontStyle = FontStyle.Normal,
-                            fontFamily = FontFamily(Font(R.font.outfit_medium))
-                        ),
-                        color = Color.Black,
-                        textAlign = TextAlign.Center
-                    )
-                }
-                Button(
-                    onClick = { deleteImage() },
-                    colors = ButtonDefaults.buttonColors(BlueSecondary)
-                ) {
-                    Text(
-                        text = "Delete Image",
-                        modifier = Modifier.fillMaxWidth(),
-                        style = TextStyle(
-                            fontSize = 18.sp,
-                            fontWeight = FontWeight.Normal,
-                            fontStyle = FontStyle.Normal,
-                            fontFamily = FontFamily(Font(R.font.outfit_medium))
-                        ),
-                        color = Color.Black,
-                        textAlign = TextAlign.Center
-                    )
-                }
-            }
-        }
-    }
-}
