+++ core/network/build.gradle.kts
@@ -89,6 +89,6 @@ dependencies {
     implementation("com.facebook.stetho:stetho:1.3.1")
     implementation("com.facebook.stetho:stetho-okhttp3:1.3.1")
 
-    //glide dependency
-    implementation("com.github.bumptech.glide:glide:4.15.1")
+    // coil
+    implementation("io.coil-kt:coil-compose:2.5.0")
 }
